Include missing factors of (mN/mX)^2 in rate_genNR.
The factors arise from the (q/mX)^2 expansion of the operator
coefficients (see eq. 6 in rate_calculators.pdf), where mX is the
DM mass. Ultimately, all factors of q^2 are assumed to be
accompanied by 1/mN^2, where mN is the mass of a nucleon, so a
factor of (mN/mX)^2 must be included with the coefficients to
obtain 1/mX^2 at the end of the calculation.

There were only a few operators that were missing these factors.